+++ src/bbochat/forms/frm_report.py
@@ -1,17 +1,17 @@
 """ReportFrame for BBO Chat."""
 
-from pathlib import Path
 import re
 import tkinter as tk
+from pathlib import Path
 from tkinter import ttk
+
+from psiutils.buttons import ButtonFrame
+from psiutils.constants import PAD
+from psiutils.utilities import window_resize
 from tkinterweb import HtmlFrame
 
-from psiutils.constants import PAD
-from psiutils.buttons import ButtonFrame
-from psiutils.utilities import window_resize
-
-from bbochat.constants import APP_TITLE
 from bbochat.config import get_config
+from bbochat.constants import APP_TITLE
 from bbochat.utilities_bbochat import display_html
 
 FRAME_TITLE = f"{APP_TITLE} - Report"