+++ HISTORY.md
@@ -1,5 +1,10 @@
 # BBO Chat - History
 
-## Version 0.0.0
+## Version 0.0.1 12 July 2026
+1. Changed data_server to data_store
+2. changed parent.parent. data_server data_store
+3. Added properties for name_1, name_2, username_1, username_2
+4. Used config.subscibe to pass colours through to forms
 
-12 Mar 2025 Initial
+
+## Version 0.0.0 12 Mar 2025 Initial